refactor(abw): separate main homepage orchestration
